Improved copy_base.py

+++ copy_base.py
@@ -4,6 +4,6 @@ import shutil
 sourceFolder = r'..\WirelessBase\src\base'
 destinationFolder = r'.\src\base'
 
-if os.path.exists(sourceFolder) and os.path.abspath(sourceFolder) != os.path.abspath(destinationFolder):
+if not os.path.exists(r'.\src\WirelessBase.h') and os.path.exists(sourceFolder) and os.path.abspath(sourceFolder).lower() != os.path.abspath(destinationFolder).lower():
     shutil.rmtree(destinationFolder,True)
     shutil.copytree(sourceFolder,destinationFolder)
